# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant QR Table Ordering System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  timestamp?: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  manager: "manager",
  chef: "chef",
  cashier: "cashier",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type TableStatus = (typeof TableStatus)[keyof typeof TableStatus];

export const TableStatus = {
  available: "available",
  occupied: "occupied",
  reserved: "reserved",
} as const;

export interface Table {
  id: number;
  number: number;
  name: string;
  capacity: number;
  status: TableStatus;
  qrToken?: string;
}

export interface CreateTableRequest {
  number: number;
  name: string;
  capacity: number;
}

export interface QrResponse {
  tableId: number;
  tableName: string;
  token: string;
  url: string;
}

export interface SessionInfo {
  tableId: number;
  tableName: string;
  tableNumber: number;
  valid: boolean;
}

export interface MenuItem {
  id: number;
  categoryId: number;
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  available: boolean;
  preparationTime?: number;
}

export interface MenuCategory {
  id: number;
  name: string;
  description?: string;
  sortOrder: number;
  items?: MenuItem[];
}

export interface CreateMenuCategoryRequest {
  name: string;
  description?: string;
  sortOrder?: number;
}

export interface CreateMenuItemRequest {
  categoryId: number;
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  available?: boolean;
  preparationTime?: number;
}

export interface UpdateMenuItemRequest {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  available?: boolean;
  preparationTime?: number;
}

export interface OrderItem {
  id: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
  notes?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  served: "served",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  tableId: number;
  tableName: string;
  tableToken: string;
  status: OrderStatus;
  items: OrderItem[];
  totalAmount: number;
  notes?: string;
  createdAt: string;
  updatedAt?: string;
}

export type CreateOrderRequestItemsItem = {
  menuItemId: number;
  quantity: number;
  notes?: string;
};

export interface CreateOrderRequest {
  tableToken: string;
  items: CreateOrderRequestItemsItem[];
  notes?: string;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  served: "served",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export type ReportSummaryOrdersByStatus = { [key: string]: number };

export type ReportSummaryTopItemsItem = {
  name?: string;
  quantity?: number;
  revenue?: number;
};

export interface ReportSummary {
  totalOrders: number;
  totalRevenue: number;
  averageOrderValue: number;
  ordersByStatus?: ReportSummaryOrdersByStatus;
  topItems?: ReportSummaryTopItemsItem[];
}

export type ValidateSessionParams = {
  token: string;
};

export type GetOrdersParams = {
  status?: GetOrdersStatus;
  tableId?: number;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  served: "served",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export type GetReportSummaryParams = {
  from?: string;
  to?: string;
};